[#server] fix bug in script

+++ server/setup.sh
@@ -13,7 +13,7 @@ PIP_PATH="$HOME_DIR/.pip"
 GITHUB_REPO_PATH="https://raw.githubusercontent.com/cxzhou35/dotfiles/main/server"
 DOTFILES=(".zshrc" ".vimrc" ".tmux.conf" ".condarc" ".gitconfig")
 
-LINK_DIR="/mnt/remote/D005/home/zhouchenxu"
+LINK_DIR="/mnt/data/home/zhouchenxu"
 
 create_dir() {
   if [ ! -d "$1" ]; then
@@ -46,30 +46,34 @@ create_dir "$LINK_DIR"
 cd $HOME_DIR
 
 # install zsh plugins
-echo -e "${GREEN}Install omz plugins..."
+echo -e "${GREEN}===== Install omz plugins ====="
 git clone https://github.com/zsh-users/zsh-syntax-highlighting.git ${ZSH_CUSTOM:-~/.oh-my-zsh/custom}/plugins/zsh-syntax-highlighting
 git clone https://github.com/zsh-users/zsh-autosuggestions ${ZSH_CUSTOM:-~/.oh-my-zsh/custom}/plugins/zsh-autosuggestions
 git clone https://github.com/paulirish/git-open.git ${ZSH_CUSTOM:-~/.oh-my-zsh/custom}/plugins/git-open
 
 # download and write dotfiles
-echo -e "${GREEN}Download and write dotfiles..."
+echo -e "${GREEN}===== Download and write dotfiles ====="
 for dotfile in "${DOTFILES[@]}"; do
-  wget "$GITHUB_REPO_PATH/$dotfile" -O "$TMP_DIR/$dotfile"
-  # write content to dotfiles
-  cat "$TMP_DIR/$dotfile" >"$HOME_DIR/$dotfile"
+  if [ -e "$HOME_DIR/$dotfile" ]; then
+    echo -e "${RED}Skip: $dotfile already exists${NC}"
+  else
+    wget "$GITHUB_REPO_PATH/$dotfile" -O "$TMP_DIR/$dotfile"
+    # write content to dotfiles
+    cat "$TMP_DIR/$dotfile" >"$HOME_DIR/$dotfile"
+  fi
 done
 
 source ${HOME_DIR}/.zshrc
 
 # create target directories and soft links
-echo -e "${GREEN}Create target directories and soft links..."
+echo -e "${GREEN}===== Create target directories and soft links ====="
 for target in "${TARGET_DIRS[@]}"; do
   create_dir "$LINK_DIR/$target"
   create_symlink "$LINK_DIR/$target" "$HOME_DIR/$target"
 done
 
 # set pip mirror(zju)
-echo -e "${GREEN}Set pip mirror..."
+echo -e "${GREEN}===== Set pip mirror ====="
 if [ ! -d "$PIP_PATH" ]; then
   create_dir "$PIP_PATH"
 fi
@@ -80,11 +84,11 @@ if [ ! -d "$LINK_DIR/miniconda3" ]; then
   echo -e "${RED}Error: Please download miniconda3 and put it in $LINK_DIR${NC}"
   exit 1
 else
-  echo -e "${GREEN}Install miniconda3..."
-  wget https://repo.anaconda.com/miniconda/Miniconda3-latest-Linux-x86_64.sh -O $LINK_DIR/miniconda3/miniconda.sh
-  bash $LINK_DIR/miniconda3/miniconda.sh -b -u -p $LINK_DIR/miniconda3
-  rm -rf $LINK_DIR/miniconda3/miniconda.sh
-  $LINK_DIR/miniconda3/bin/conda init zsh
+  echo -e "${GREEN}===== Install miniconda3 ====="
+  wget https://repo.anaconda.com/miniconda/Miniconda3-latest-Linux-x86_64.sh -O $HOME_DIR/miniconda3/miniconda.sh
+  bash $HOME_DIR/miniconda3/miniconda.sh -b -u -p $HOME_DIR/miniconda3
+  rm -rf $HOME_DIR/miniconda3/miniconda.sh
+  $HOME_DIR/miniconda3/bin/conda init zsh
 fi
 
 rm -rf "$TMP_DIR"
